fix: split AuthService validators into AuthQueryService for CQS alignment
AuthService is the command-side service for the auth domain but was
hosting two pure validators (checkPassword, checkAuthNum) that read
state without mutating any domain object. Per the project's CQS rule
(Service = command, QueryService = query/check/exists), move both
validators into a new AuthQueryService and update AuthUseCase to call
the query service for them.

- New: domain/auth/application/query/AuthQueryService
    - checkPassword(rawPassword, hashedPassword)
    - checkAuthNum(email, authCode)
- AuthService: drop the two methods + the now-unused PasswordEncoder
  field; keep generateToken / refreshToken / createRandomNum
- AuthUseCase: inject AuthQueryService and route both check calls
  through it

checkAuthNum still does its 'delete the one-time code on success' write
to Redis — this is the atomic cleanup tied to the validation, not a
separate command, so keeping it inside the check method preserves the
single-call semantics that the use case relies on.
